~ 6.0 Cleaning JS

+++ src/components/Nweet.js
@@ -8,7 +8,9 @@ const Nweet = ({nweetObj, isOwner}) => {
         const ok = window.confirm("Are you sure you want to delete this nweet?");
         if(ok) {
             await dbService.doc(`nweets/${nweetObj.id}`).delete();
-            await storageService.refFromURL(nweetObj.attachmentUrl).delete();
+            if(nweetObj.attachmentUrl != null) {
+                await storageService.refFromURL(nweetObj.attachmentUrl).delete();
+            }
         }
     }
     const toggleEditing = ()=> setEditing(prev => !prev);
